+++ src/main/java/ru/yandex/practicum/filmorate/dao/event/EventDao.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.dao;
+package ru.yandex.practicum.filmorate.dao.event;
 
 import ru.yandex.practicum.filmorate.model.Event;
